requirements.txt
# Core FastAPI dependencies
fastapi==0.116.1
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Pydantic for data validation
pydantic==2.8.2

# Database
sqlite3

# Environment variables
python-dotenv==1.0.1

# LangChain and LangGraph
langchain-core==0.2.31
langchain-groq==0.1.9
langgraph==0.2.3

# Audio for alarm system
playsound==1.3.0

# Image processing
Pillow==10.4.0

# HTTP client
httpx==0.27.0

# Additional utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
